// src/types/BorderAccent.ts
export type BorderAccent = 
  { from: string; to: string } // градиент из двух цветов
  | 'yellow' // желтый цвет
  | 'red' // красный цвет
  | 'green' // зеленый цвет
  | 'blue' // синий цвет
  | 'purple' // фиолетовый цвет
  | 'pink' // розовый цвет
  | 'cyan' // голубой цвет
  | 'orange' // оранжевый цвет
  | 'white' // белый цвет
  | 'gray' // серый цвет
  | 'black'; // черный цвет